pyproject.toml
[tool.poetry]
name = "datagaze-api"
version = "0.1.0"
description = ""
authors = ["Abduraxmonnn <abduraxmonasatullayev35@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.7.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
